requirements.txt
asgiref==3.5.0
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.10.0
blog==1.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-grappelli==3.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
docopt==0.6.2
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
legacy==0.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.11.2
pymdown-extensions==9.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
whitenoise==6.0.0
